Refactor charge consumption on item activation (#60039)
* activation_consume

* consume amount

* activation_charges_sufficient

* return values

* charge consumption

* some test fixes

* test fixes

* astyle

* manhacks

* place_tra

* bionic and furniture

* int ammo_scale

* cleanup

* cleanup

* cleanup

* fix comparison

* dont access charges_to_use directly

* unused arguments

* unused arguments

* for nontools

* use ammo_consume for all

* mod bots

* fix flag

* use ammo_sufficient

* use ammo_sufficient

* no need to check for item deletion

* astyle

* default charges requried

* revert unneeded

* revert unneeded

* update comment

* move to cpp

* note about deletion

+++ src/activity_actor.cpp
@@ -1003,7 +1003,7 @@ void hacksaw_activity_actor::do_turn( player_activity &/*act*/, Character &who )
 
     if( tool->ammo_sufficient( &who, method ) ) {
         int ammo_consumed = tool->ammo_required();
-        std::map<std::string, float>::const_iterator iter = tool->type->ammo_scale.find( method );
+        std::map<std::string, int>::const_iterator iter = tool->type->ammo_scale.find( method );
         if( iter != tool->type->ammo_scale.end() ) {
             ammo_consumed *= iter->second;
         }
@@ -4855,7 +4855,7 @@ void prying_activity_actor::do_turn( player_activity &/*act*/, Character &who )
 
     if( tool->ammo_sufficient( &who, method ) ) {
         int ammo_consumed = tool->ammo_required();
-        std::map<std::string, float>::const_iterator iter = tool->type->ammo_scale.find( method );
+        std::map<std::string, int>::const_iterator iter = tool->type->ammo_scale.find( method );
         if( iter != tool->type->ammo_scale.end() ) {
             ammo_consumed *= iter->second;
         }
